app/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
python-multipart
requests
